Move informer factor creation into controller package
