+++ internal/controller/scheduler/scheduler.go
@@ -18,9 +18,6 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/selection"
-	"k8s.io/client-go/informers"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/utils/pointer"
 )
@@ -43,31 +40,6 @@ func New(logger *zap.Logger, client kubernetes.Interface, cfg Config) *worker {
 		client: client,
 		logger: logger.Named("worker"),
 	}
-}
-
-// returns an informer factory configured to watch resources (pods, jobs) created by the scheduler
-func NewInformerFactory(
-	k8s kubernetes.Interface,
-	namespace string,
-	tags []string,
-) (informers.SharedInformerFactory, error) {
-	hasTag, err := labels.NewRequirement(config.TagLabel, selection.In, config.TagsToLabels(tags))
-	if err != nil {
-		return nil, fmt.Errorf("failed to build tag label selector for job manager: %w", err)
-	}
-	hasUUID, err := labels.NewRequirement(config.UUIDLabel, selection.Exists, nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed to build uuid label selector for job manager: %w", err)
-	}
-	factory := informers.NewSharedInformerFactoryWithOptions(
-		k8s,
-		0,
-		informers.WithNamespace(namespace),
-		informers.WithTweakListOptions(func(opt *metav1.ListOptions) {
-			opt.LabelSelector = labels.NewSelector().Add(*hasTag, *hasUUID).String()
-		}),
-	)
-	return factory, nil
 }
 
 type KubernetesPlugin struct {
